# packages/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first AI agent observability \u2014 ingest and query API",
    "title": "Korveo API",
    "version": "0.6.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/openai/{path}": {
      "post": {
        "description": "Forward an OpenAI-compatible request to the configured upstream\nand capture a Korveo span on the way back.\n\nPath is preserved verbatim under the upstream base, so a client\npointed at ``http://localhost:8000/v1/openai`` sees the upstream\nsurface unchanged: ``/v1/chat/completions``, ``/v1/embeddings``,\n``/v1/models``, etc. all work.\n\nStreaming responses (``stream: true``) are passed through chunk\nby chunk; we buffer in parallel to assemble the span at the end.\nNon-streaming responses are returned in a single Response.\n\nErrors from the upstream are returned as-is; a span is still\nemitted so operators can see failed calls in the dashboard.",
        "operationId": "openai_proxy_v1_openai__path__post",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-korveo-project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Korveo-Project"
            }
          },
          {
            "in": "header",
            "name": "x-korveo-upstream",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Korveo-Upstream"
            }
          },
          {
            "in": "header",
            "name": "x-korveo-session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Korveo-Session"
            }
          },
          {
            "in": "header",
            "name": "traceparent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Traceparent"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Openai Proxy"
      }
    },
...
  }
}